+++ test/bottom_bar_test.dart
@@ -1,11 +1,9 @@
 import 'package:MobileOne/localization/delegate.dart';
 import 'package:MobileOne/localization/supported.dart';
 import 'package:MobileOne/pages/Mainpage.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
 
 Widget buildTestableWidget(Widget widget) {
   return MaterialApp(
@@ -31,10 +29,6 @@ Widget buildTestableWidget(Widget widget) {
       ),
       home: widget);
 }
-
-class FirebaseUserMock extends Mock implements FirebaseUser {}
-
-class FirebaseAuthMock extends Mock implements FirebaseAuth {}
 
 void main() {
   setSupportedLocales([Locale('fr', 'FR')]);